pyproject.toml
[tool.poetry]
name = "kotaztempaizer"
version = "0.1.0"
description = "Kotaz Tempaizer - шаблонизатор написанный лично для моих нужд"
authors = ["Kotaz <semechkagent@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
GitPython = "^3.1.24"

[tool.poetry.dev-dependencies]
pyment = "^0.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
